Remove static usage of kMaxThreads (#642)
This PR removes the fixed `kMaxThreads` dependency from the transport
loop and makes the relevant counters runtime-sized from the Geant4
worker count.

This PR should fix the nightlies, which ran out of stack in debug mode
when the kMaxThreads was increased to 512 to accommodate some larger
runs on H100 GPUs.

The fixed-size per-event arrays in `Stats` are replaced with device
pointers sized at GPU initialization.
Thereby, the population counting and reset kernels are updated to use
`numThreads` and dynamic shared memory.
Note that now the flags for the FinishOnCPU must be copied and cannot be
passed as a fixed size array anymore. This copy should be negligible.

+++ include/AdePT/transport/AdePTTransport.hh
@@ -35,7 +35,7 @@ public:
   uint64_t fAdePTSeed = 1234567;
 
 private:
-  unsigned short fNThread{0};             ///< Number of G4 workers
+  unsigned int fNThread{0};               ///< Number of G4 workers
   unsigned int fTrackCapacity{0};         ///< Number of track slots to allocate on device
   unsigned int fStepCapacity{0};          ///< Number of step slots to allocate on device
   int fDebugLevel{0};                     ///< Debug level
